New Worm Model, works
Worm model, refactor of machine file to multiple machine files

Thunder/lightning readded, has been instead used to to the targeting for the worm. worm now spawns where lightning hits
