+++ prototypes/resources.lua
@@ -172,7 +172,9 @@ data:extend({
     stateless_visualisation = nil;
     map_color = {252, 255, 39},
     map_grid = false
-  --[[},
+  }
+})
+    --[[
   {
     type = "resource",
     name = "multi-ore",
