+++ scripts/init.lua
@@ -1,9 +1,9 @@
 script.on_init(function()
-    local water_condensers = {}
+    --Placeholder
 end)
   
 script.on_configuration_changed(function()
-    local water_condensers = {}
+    --Placeholder
 end)
